Added offsets for glow esp. Added realization for esp itself. Added methods for health bars but not used in a hack (looks kinda ugly and dull). In exchange for health bar added color adjustment for a health value (from green to red).

+++ CSAPP/MemScan.cpp
@@ -70,7 +70,7 @@ ProcessHandler::~ProcessHandler()
 	CloseHandle(hProc);
 }
 
-DWORD ProcessHandler::GetProcID() const
+uintptr_t ProcessHandler::GetProcID() const
 {
 	// returning process ID
 	return procID;
